p3/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
numpy==2.1.1
pandas==2.2.3
protobuf==5.27.2
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
